// temp-job-frontend/features/advertisement/schemas/advertisement.schema.ts
import { z } from "zod";

export const AdvertisementSchema = z.object({
  id: z.number(),
  pathname_to_display: z
    .string()
    .min(1, { message: "กรุณาระบุ url ที่ต้องการแสดงโฆษณา" }),
  created_at: z.string(),
  image_url: z.string(),
});

export type Advertisement = z.infer<typeof AdvertisementSchema>;